Allow a single register to contain multiple instruction values.
Previously, we had a 1:1 mapping for `register:instruction`; this commit
changes this ratio to 1:n.

The main use case for this is sign/zero extension. In many cases it's a
no-op. Consider this:

```
%0: i32 = param reg
%1: i64 = zext %0
```

`%0` has `Undefined` upper bits; in `%1` those bits are `ZeroExtended`;
but both `%0` and `%1` in a sense "share" the same "set bits". Before
this commit we'd have generated code along these lines:

```
; %0: i32 = param reg
; %1: i64 = zext %0
mov ebx, eax
```

so `%0` and `%1` would reside in two registers. After this commit, we
generate code roughly along these lines:

```
; %0: i32 = param reg
; %1: i64 = zext %0
mov eax, eax
```

and both `%0` and `%1` will now "reside" in `rax`.

It's probable that this commit will give us a small speed boost, because
it will reduce register pressure. That, however, is not really the main
reason for this commit: it's really a (fairly easily testable) warmup
for removing dealiasing at the start of a trace.