Remove the "meaningful register" hack.
Our register allocator used to be very bad when register pressure was
high and when we didn't have peeling. The hack this commit removes thus
makes a smaller difference to performance than it once did -- and it
also removes a source of incorrectness (in the same sense as the
dealiasing removed in the previous commit).