We need to tell the register allocator which registers to keep.
When the register allocator is shuffling registers around, it needs to
know which registers need to keep their value maintained. The `if dst ==
src` check elided that information, which meant that the register
allocator could -- if you were very unlucky! -- use an active register
as a temporary register. Simply removing the check allows the register
allocator to do the right thing.