Deal with tracing starting in nested interpreter frames.
The "main" test here is `trace_while_executing1.c`, which models a
failure we see occasionally in yklua: a JIT frame ends up calling an
interpreter frame, which starts tracing, and tracing doesn't falls back
into the JIT frame, which then happens to deopt. It took me quite a
while to convince myself that's what happening, but it really is!

`trace_while_executing2.c` is a sort-of variant test of something
similar that we happened to already deal with correctly, but which no
test captured.