Implement support for threads.
There wasn't support for threads in yk as it requires the shadow stack
to be a thread local. This was previously not possible due to us using
LLVM to compile our traces. Now that we use our own JIT compiler, there
isn't much standing in the way of thread support as this commit shows.

The first (and easier) step is to change the shadow stack to a thread
local in ykllvm. We then intercept the creation of threads so we can
allocate a new shadowstack for each thread. At JIT compile time, we
first calculate the offset of the thread local in relation to the
segment register `fs`. At JIT runtime (when the segment register has
been updated by a new thread) we can then use this offset to calculate
the thread local's pointer.