Flag "tracing-aborted" as a warning.
Some "tracing-aborted" cases were printed as a warning, but most were
printed as `JITEvent`s. Since it's increasingly clear that aborted
traces are something we need to address, hiding them under endless
"executing trace" (etc.) messages feels unwise.

This commit consistently prints "tracing-aborted" as a warning, which is
a baby step in helping us why and when some traces are aborted.