Give `icmp_guard` a better way of assigning registers.
Previously we allocated registers for both the `icmp` and the `guard` in
one go, but that causes unspilling issues (and also caused us to
generate slightly worse code than necessary, because we needed 3
registers in a combined icmp/guard rather than than 2 and then 1
registers separately). To work around that, however, seems tricky,
because `assign_reg`s can set CPU flags.

This commit offers an alternative mechanism (which, currently, we don't
need, but we probably will soon so it's `todo`ed) where in the worst
case we will need to `push`/`pop` if we encounter a non-zero-extended
`i1` that we need to spill.