Give guards a better way of assigning registers.
Previously we allocated registers for guards (in both their "normal" and
"icmp/guard" forms) in a way that:

  1. causes unspilling issues
  2. also caused us to generate slightly worse code than necessary (e.g.
     because we needed 3 registers in a combined icmp/guard rather than
     than 2 and then 1 registers separately).

It took me a while to understand what we really need. In essence, guards
(and -- when we come to them! -- calls with "more operands than
registers in the ABI") are different than other things we need to
allocate registers for: normally we know that if an instruction doesn't
say "I want to use this operand", we can (in a sense) discard it. But
with guards, there is a (long...) list of live variables that we don't
"use" directly, but which we can't discard.

This PR gives a very simple API in the allocator specifically for
guards. The way to think of it is "give me registers for X; but if there
are any values in there, spill them, even if I haven't said I want those
values".