Use a simple heuristic to decide which registers to try moving.
Previously this part of the code worked on the "be lucky" principle: it
tried registers in a random order to decide whether to move/spill their
values. This commit adds a very simple heuristic: we try to move (rather
than spill) those registers whose values will be used most often in the
remainder of the trace.