Tweak the null location in yk.h.
I need the ability to (from within the C code of an interpreter) say
whether a location is a null location.

I tried using the (unused, it seems) `YKLOCATION_NULL` macro in `yk.h`,
but it's not valid C. You can't cast an integer to a struct.

This change replaces it with an inline function that does a null
location check by comparing the state field to zero. I've done it this
way to avoid calling into ykcapi.so, which would come at a higher cost.

This is required for an upcoming change to fix placement of YkLocations
in yklua.