Connector traces.
Connector traces are non-looping traces that end by jumping to an
existing `CompiledTrace`. They allow us to "stay within machine code"
much more often: not only is that faster than the interpreter, but it
also means that we sidestep deopt costs. RPython calls this kind of
traces "bridges", but for some reason that name never gave me the right
intuition, so I'm tentatively trying "connector" instead (I've
previously tried "link", and "jump_to", but those weren't improvements).
I suggest we don't get hung up on the name yet, because the semantic and
performance change is what this commit is all about!

The most obvious use for connector traces is when you have inner/outer
loops e.g.:

```
for x in .. {
  for y in .. { }
}
```

In most cases the inner loop will get hot first, and lead to a
`CompiledTrace`. When we later trace the outer loop, what should we do?

  * We could unroll the inner loop, but that's nearly always a bad idea
    (we once did this!).
  * We could abort compilation (which is what we did before this
    commit).
  * We can create a connector trace from the start of the outer loop
    to the inner loop's `CompiledTrace`.

That's what this commit does -- and more!

It also alters sidetraces so that they no longer have to end in a root
trace. In other words, sidetraces can (optionally) be a kind of
connector trace. In the example above, we will end up with a "connector
sidetrace" from the end of the inner loop to the outerloop's connector
trace.

It also makes all of this play well with parallel compilation:
compilation jobs can now have a dependency on a yet-to-be-compiled trace
(hence the `CompiledTraceId` -> `TraceId` renaming). This means that we
can trace the outer loop while the inner trace is still compiling, and
put the outer trace on the compilation queue: we won't start compiling
it until the inner trace has been compiled.

This commit doesn't tie up every possible loose end (e.g. we can now end
up with "dead" compilation jobs if a compiling trace fails to compile
and there are dependent traces waiting upon it), but it's already
sufficiently big and complex that I think trying to do more will be
overwhelming.

This commit doesn't help every benchmark, but when it does, it can
be very helpful e.g.: deltablue is improved by a smidgen under 2x;
richards by 1.15x. On deltablue this is because we previously compiled
about 13 traces successfully and aborted 48: with this commit we
successfully compile 104 traces and abort none. On richards we also
successfully compile many more traces, but deopt time goes down by
10x.