Propagate heap knowledge of constants across peeling.
If we know that, say, `%42 + 8` at the end of the header is the constant
`8` at the end of the header, we can propagate that knowledge to the
body. This has the nice property that the body can later store a
different constant to that location (but, by definition, it must store
`8` back to that location by the end of the trace!), and the right thing
will happen.

Across a few tight loop benchmarks I tried this gives a 3-5% boost.