Only propagate heap information for unchanging pointers.
The very simple new test captures something that I didn't get right
before. In essence, it's a "zero all memory" trace but before we didn't
notice that the pointer increments on each iteration of the trace.

This commit checks that we only propagate knowledge of pointers that are
truly loop invariant. We do that by checking that the same SSA value
that was an input to the trace header is passed to the trace end.