Fix idempotent promotes.
We were immediately consuming the return value from the promote buffer
upon encountering a call to an idempotent function. If the callee calls
functions that then promote their own values, then this means we consume
the promote buffer in the wrong order.

This change ensures we only consume the return value from the promote
buffer *after* we have returned from the idempotent function, and thus
all inner promotions will already have been handled.

We achieve this by introducing a "pending" state for idempotent
constants. When the trace builder encounters a call to an idempotent
function it sets the `idem_const` field to `IdemConst::Pending`. When
the trace builder has finishes outlining, it checks if it just emitted a
call instruction with a pending `idem_const`, and if so it replaces it
with `IdemConst::Const(cidx)`.