Deal correctly when peeling discovers guards that always fail.
When we're peeling we might discover that the peeled iteration can never
execute the same as the first iteration. Consider a loop such as:

```
for x in 0..100 {
  if x % 2 == 0 { ... }
  else { ... }
}
```

In other words, successive iterations of the loop flip between the true/else
branches. Thus, whatever iteration we trace, the peeled iteration will never
be able to execute to its end: it will always fail at a guard before that
point.

When we're optimising the peeled iteration, our analyses can thus detect
guards that will always fail. As soon as we discover one of those, there's
no point trying to optimise the guard or anything after it in the trace: the
guard will always fail at run-time.

This commit thus fixes a panic first discovered in
https://github.com/ykjit/yk/issues/1689.