Replace `YKD_LOG_IR=trace-kind` with `debugstrs`.
This allows you to print a trace with only its debug strings. This is a
simple hack for working out what opcodes were traced.

As a result of this, I'm not sure `trace-kind` has any use, so I've
ripped it out. If we do need it, readding it is trivial.