Fix yklua submodule uninitialised error message.
The error message that informs us if the yklua submodule hasn't been
initialised yet isn't showing. The condition checks if the `tests/yklua`
folder exists, which is always true. We need to instead check whether it
contains any files to determine if the submodule has been initialised.

Which file to check for is a bit tricky as the yklua repo might change
in the future, but checking for the `Makefile` is probably good enough
for now. The proper fix would be to use `git submodule status` but that
change is a bit more involved.