Don't bork on non-UTF-8 debug strings.
yklua can produce non-UTF-8 debug strings (Lua source files are not
required to be UTF-8). We could handle this (albeit awkwardly) in yklua,
but it's a general problem: C strings aren't required to be UTF-8. So
rather than forcing the interpreter to deal with this, use a lossy UTF-8
conversion.