Reenable the `OutsideYk` timing state.
When using `YKD_STATS` I noticed that the `OutsideYk` time wasn't
reporting correctly. This was because this state was only set after
deopting, but not after tracing, compiling, etc. Setting this state at
the end of `StopTracing`, `Stop-Side-Tracing`, and `AbortTracing`, just
before we return to the interpreter, will measure the time we spend
outside of yk correctly again.