Avoid spilling variables used in body/header end twice.
Some variables in the body/header end have to end up on the stack for
the trace we're jumping/looping to. Previously `write_jump_vars` handled
the situation where we'd already spilled a variable somewhere and needed
it to be in a different place.

This commit avoids this dance by tracking which variables will be
spilled at the trace end and putting them in the right place straight
away.

This can make quite a difference at the level of an individual trace.
e.g. here's a diff before/after of one part:

```diff
 ; sidetrace_end 0x7abcdef [%0, %1, %2, %3, %4, %5, %6, %7, %8, %9, %10, %11, %12, %13, %14, %15, %16, %17, %18, %19, %20, %21, %22, %23, %24, %25, %26, %27, %28, %29, %30, %31, %32, %102, %104, %105, %108, %109, %115, %116, %117, %118, %114, 0i32, 0x7abcdef, 16941972i32, 0x7abcdef]
 mov [rbp-0x7abcdef], rax
 mov [rbp-0x7abcdef], rsi
-mov rcx, [rbp-0x7abcdef]
-mov [rbp-0x7abcdef], rcx
-mov rcx, [rbp-0x7abcdef]
-mov [rbp-0x7abcdef], rcx
-mov [rbp-0x7abcdef], r11
-mov [rbp-0x7abcdef], r8
+mov [rbp-0x7abcdef], rdx
 mov [rbp-0x7abcdef], r10
-mov rcx, [rbp-0x7abcdef]
-mov [rbp-0x7abcdef], rcx
-mov [rbp-0x7abcdef], rdi
+mov [rbp-0x7abcdef], r8
 mov dword ptr [rbp-0x7abcdef], 0x7abcdef
 mov rcx, 0x7abcdef
 mov ebx, 0x7abcdef
```

This is probably most effective, though, on headerandbody traces, though
those tend to do less spilling so we tend to only win perhaps one or two
spills on average there.