Align jump targets inside traces.
This change aligns the jump targets of the trace loop and side traces jumping
back into a parent trace to 16 bytes in order to improve or at the very least
stabilise performance.

Before this, changes as small as a single extra instruction inside the trace
header, had the potential to significantly alter the performance of the trace,
making it difficult to test changes in for example yklua.