Fix the "skip the initial block" logic in the trace builder.
The way the pattern is matched, assumes that the block mapped to exactly
one IR block, but this appears to not be the case.

Whereas before I could get this to crash relatively quickly, now it
takes much longer, and when it does crash, it's for a different reason:

```
thread '<unnamed>' panicked at hwtracer/src/pt/ykpt/mod.rs:449:13:
internal error: entered unreachable code: compressed returns are disabled in collect.c
```

(this after 245 try_repeats)

As it happens, I'm already investigating this crash, as it's the same
one that stops us from not emitting stackmaps into outlined functions.
I *think* this is a separate issue.