Handle synthetic block terminators.
LLVM has the concept of "synthetic block terminators", which are
composite terminators constructed of 2 (or in theory more) conditional
branch instructions.

Examples are:
  - COND_NE_OR_P
  - COND_E_AND_NP

When doing compiler-assisted PT decoding, we were assuming that all
blocks with a condition successor would consume exactly one TNT bit,
which in light of the above isn't true.

If a block is terminated by two conditional branches then we will have
to consume between 1 and 2 TNT bits, depending upon whether the first
branch was taken or not.

This fixes a long-standing PT bug where the PT decoder would crash
expecting a TIP packet from an uncompressed return. We were consuming one
TNT where we needed to consume two, meaning that the otherwise
unconsumed TNT was being used to decide if a return was compressed or
not. Hilarity ensues.

This also clears some other long-standing confusion about why we can't
throw away decoder state across a PSB+. With the above fix applied, we
can!