When updating the tail value in ring buffers, write the unwrapped value.
When collecting a PT trace we have to read data out of mmaped ring
buffers at regular intervals.

Head and tail markers are shared between userspace and the kernel. The
kernel updates the head pointer to mark new data userspace hasn't seen
yet, and userspace updates the tail to tell the kernel what we've read
up until.

The kernel expresses the head and tail markers as unsigned 64-bit values
which don't wrap at the extent of the size of the ring buffer. In order
to interpret these markers, userspace is expected to wrap (modulo) the
markers with the size of the ring buffer.

Before this change we were updating the tail marker with a wrapped
value, which I think was corrupting the kernel's view of reality once
the AUX buffer wrapped for the first time. After this point we would get
zero-length PT traces devoid of any PT packets and were unable to ever
trace properly again: we'd get lots of "premature end of trace" errors
(e.g. with havlak).

This change writes back the unwrapped tail value instead. With this
change, we get proper PT traces after the AUX buffer wraps and the
"premature end of trace" errors are gone.

This is likely the cause of the long-standing "cached PT stuff makes
weird things happen" because when you re-use the PT fd, you are more
likely to see the AUX buffer wrap.