Use a dedicated thread local to detect if we're tracing.
When a VM is interpreting, it still has to call `yk_promote`,
`control_point` and so on -- which then check if we're tracing and do
(or don't...) do something as appropriate. Currently that check is
fairly expensive: we do a `borrow` of a `RefCell`, then look at the
contents of the last element in a list.

This commit adds a new thread local `THREAD_IS_TRACING` which is an
`AtomicBool` which allows us to do a simple load/compare/branch.
Benchmarking this beyond a shadow of doubt is hard because it's during a
part of the system that can be quite non-determinstic. However, it looks
to me like a small win, perhaps 1-2%. This is highly dependent on how
much time is spent in the interpreter: if, for example, I turn
synchronised compilation on for the `json` benchmark, I can observe a 5%
difference, but with synchronised compilation on, there's a lot of
noise.