Only wake up other threads if necessary.
When we've compiled a root trace, other jobs waiting on it can be
processed. Previously we blithely woke up one of them.

This commit checks how many jobs are waiting on the just-compiled
thread, and tries to wake up an appropriate number of threads. In most
cases, there are no jobs (so no waking to do!), or 1 job (the same
situation as before). Sometimes, though, there are multiple jobs
waiting, in which case we notify every worker thread. In other words,
when there's no work to do, we now do less work; when there's lots of
work to do, we do it straight away. This reduces latency somewhat,
because previously some jobs had to wait around for a worker thread to
realise there was work to do.