Create trace IDs at tracing time.
Previously we created new trace IDs when we were about to compile a
trace. This commit changes this so that we create new trace IDs as we
start tracing. This has a couple of direct advantages:

  1. We can now check the "is this the location we started tracing?" by
     trace ID rather than `Arc::ptr_eq`, which always felt... icky.
  2. We can now detect tracing in another thread and create connector
     traces to it.

It has an indirect advantage that we will exploit in an upcoming commit:
it will allow us to put jobs in the queue that "end" in traces that have
not yet been created.