Split apart loop and guard tracing transition.
Intermingling these two had become very tricky, and I was struggling to
tell when we were handling one case over another.

Splitting them apart also showed that we were not handling one case
sensibly: when we encounter a `Location` that is (or was!)
`SideTracing`, a guard trace can stop. This allows us to delete a
now-bogus test.