Make guard traces end at the first non-null-`Location` they encounter.
Previously, guard traces kept tracing until they encountered a compiled
/ compiling / tracing `Location`. Most of the time, this would happen at
the first point they encountered a non-null-`Location`, but sometimes
they would happily breeze through several `Location`s that were
`Counting` before finding one to end at.

This commit changes that: guard traces now always terminate at the first
non-null-`Location` they encounter. If that is a `Counting` `Location`,
then we immediately start tracing that `Location` and add the new trace
later as a dependency of the guard trace. That means that we
put the guard trace in the job queue before we've even started tracing
the new `Location`! Thus from the perspective of trace output, the
`Counting` location finishes compilation first, even though it was
traced second. Consider this output from `nested_loops.lua`:

```
--- Begin debugstrs: connector: nested_loops.lua:44: ADDI ---
  nested_loops.lua:45: LOADI
  nested_loops.lua:45: LOADI
  nested_loops.lua:45: LOADI
  nested_loops.lua:45: FORPREP
  nested_loops.lua:46: ADDI
--- End debugstrs ---
--- Begin debugstrs: side-trace: nested_loops.lua:46: ADDI ---
  nested_loops.lua:43: FORLOOP
  nested_loops.lua:44: ADDI
--- End debugstrs ---
```

We traced the guard trace first, but it couldn't be compiled until the
next `Location` -- which ended up being a coupler trace -- was first
compiled.

One of the aims of this commit is to keep guard traces as small as
possible: this will tend to (slightly) reduce the number of traces we
end up creating overall. If you consider the example output above,
previously (after some aborting of compilation and retrying etc), we'd
have ended up with one guard trace covering all of that, but,
eventually, the second `Location` might also have created a coupler
trace anyway due to other control flow. That coupler trace would have
overlapped with most of the guard trace, so we'd have a largely
duplicated -- but impossible to reconcile -- trace tree.

Right now, the effect of this commit is fairly small and benchmark
dependent -- e.g. it benefits Richards by ~3%, but slows deltablue down
by ~2% -- but it opens up further possibilities for code simplification
and, eventually, more sophisticated control of what and how we trace.