Remove `HotLocation::SideTracing`.
Thanks to several preceding changes, ne no longer need this, as we can
record the necessary information in `MTThread::Tracing` (and, happily,
we no longer need to cart around `root_ctr`). There is probably a small
performance gain from doing this, as we don't need to do quite as much
locking and reading / writing of `HotLocation` kinds, but it's within
the noise.