Fix code generation with imm64s.
In most cases, x64 sign extends `imm64` even though it's represented as
an `i32`. This commit fixes the mistakes that I made when I didn't
account for this.

There is at least one further optimisation we could make to the code
below but this *probably* still generates "optimal" code in the vast
majority of cases we see, so the extra complexity is not yet obviously
worth it.