A spare register isn't always available.
Previously we had -- as the comment said -- normally been lucky in
`write_jump_vars` and not needed to spill a register. I've now
encountered situations where that isn't true and we need to spill a
register.

This commit handles that case --- but it does try very hard not to
spill! In order to do that, we have to sort-of duplicate a `match`: this
does allow us to escape a spill in a small but notable number of cases
(enough that I can just about measure it in a small benchmark).