Don't spill patch register before a guard.
Guards not a patch (or "jump") register: when we don't have a spare
register, we have to spill one. Previously we did this before a guard:

```
cmp ...
mov [rbp-...], r15
jnz ...
```

So even if you didn't take the guard, you still paid the spill (and
sometimes unspill) penalty.

This commit changes the code above to:

```
cmp ...
jnz ...
```

The spill still has to go somewhere, of course! It now goes in the deopt
code (see the test below), so you only pay the spill penalty if the
guard is taken.

This is a fairly big change: it means we have to snapshot the register
allocator at the point of a guard and restore that state at each deopt
point. The performance change is, at the moment, probably fairly small:
but this is the beginning of moving more code into guards, so this is
in a sense the "smallest" change which nudges us in that direction.