Remove `SideTraceInfo` from mt.rs.
The concept of "side trace info" is quite specific to the jitc_yk
backend, so this commit is the first step in pushing that all the way
into that backend. We now give backends sufficient information
(`parent_ctr`, `gidx`, and `target_ctr`) to store whatever they need for
guard traces in the way that's most efficient for them.