Push `SideTraceInfo` into `jitc_yk`.
Most obviously, this commit simply pushes `SideTraceInfo` -- which is
rather specific to jitc_yk -- into jitc_yk. That makes the "top-level"
parts of yk a little more abstract.

Less obviously, this commit implicitly acknowledges that jitc_yk is *de
facto* x64 only. Our pretence that it is not just makes the code hard to
read -- in several places, what look like references to traits are in
fact references to concrete structs. There is next to no chance of
unpicking this, so this commit makes jitc_yk x64 only *de jure*.

Note: these two things are related, and you can't do them in two
commits, or do the former without the latter. It does feel a bit odd to
make the code seemingly less abstract, but really we're just
acknowledging reality.