Calculate the transitively reachable set of "guard body" instructions.
This means that we now pull even "indirect" instructions into guard
bodies. Because the register allocator does quite a bad job inside guard
bodies, this isn't as big a win as the previous PR, but it's still
useful. The winners are:

```
DeltaBlue/YkLua/12000     6.72% faster
Havlak/YkLua/1500         3.83% faster
NBody/YkLua/250000        3.25% faster
```