Fix virtual address to file offset mapping.
During working on something else I realised that the way we convert
between offsets and virtual addresses is really broken.

After fixing the routines that do this stuff, I found other parts of the
system (mostly the PT decoder) that had been relying on the broken logic
and had to find and fix all of those too.

One realisation is that addresses appearing in the LLVM blockmap are not
offsets but virtual addresses. This is because the addresses are emitted
as symbols in a loadable section, meaning that they get relocated! The
aforementioned broken logic above led me to think that this was not the
case.

Performance is a mixed bag (3 executions, 5 iterations):

```
Benchmark                  Datum0 (ms)  Datum1 (ms)  Ratio  Summary
 LuLPeg/YkLua/default       1822         1678         0.92   7.89% faster
 BigLoop/YkLua/1000000000   1928         1846         0.96   4.25% faster
 HashIds/YkLua/6000         1174         1139         0.97   2.96% faster
 knucleotide/YkLua/default  1415         1375         0.97   2.84% faster
 NBody/YkLua/250000         529          517          0.98   2.19% faster
 Permute/YkLua/1000         869          855          0.98   1.65% faster
 List/YkLua/1500            882          869          0.98   1.50% faster
 Bounce/YkLua/1500          1014         1000         0.99   1.44% faster
 Json/YkLua/100             1571         1551         0.99   1.24% faster
 Heightmap/YkLua/2000       782          776          0.99   0.76% faster
 Richards/YkLua/100         4406         4379         0.99   0.61% faster
 Sieve/YkLua/3000           399          400          1.00   0.24% slower
 fannkuchredux/YkLua/10     1408         1413         1.00   0.35% slower
 spectralnorm/YkLua/1000    974          977          1.00   0.36% slower
 binarytrees/YkLua/15       1349         1357         1.01   0.66% slower
 Mandelbrot/YkLua/500       136          137          1.01   0.76% slower
 CD/YkLua/250               4132         4187         1.01   1.35% slower
 Storage/YkLua/1000         3202         3263         1.02   1.88% slower
 Havlak/YkLua/1500          8554         8758         1.02   2.39% slower
 Towers/YkLua/600           963          995          1.03   3.28% slower
 DeltaBlue/YkLua/12000      1295         1357         1.05   4.79% slower
 Queens/YkLua/1000          530          562          1.06   5.92% slower
 revcomp/YkLua/default      611          665          1.09   8.87% slower
 fasta/YkLua/500000         496          731          1.47   47.23% slower
 ```

Shame about fasta, but this change is for soundness, so we probably
should move forward anyway.

I have a follow up in mind to do a load of tidying that I hope will
result in some speedups. Let's see.