Rename `aot_ir::InstIdx` to `BBlockInstIdx`.
In JIT IR we use `InstIdx` to mean "index into the trace", so it's easy
to misread `aot_ir::InstIdx`, particularly as `jit_ir::InstIdx` is
somewhat closer (but definitely not the same as!) `aot_ir::InstId`.

This commit renames the struct to make it a little clearer what's going
on.