Rename `GuardIdx` to `GuardId`.
This struct sits outside a specific compiler, and there's no reason that
we need to force it to be an index.

I did toy with making this a `trait`. That would allow each compiler to
hide its preferred notion of "guard ID" more completely but (a) we'd
then have to `Box` guard IDs which, though infrequent, still feels a bit
wasteful (b) each compiler would have to `downcast` the result, which
always feels a bit icky.

In that sense, "guard ID as `usize`" is a pragmatic compromise. It's
efficient and flexible (a compiler could stuff a pointer in it!), even
if it doesn't quite integrate with the type checker in a way that I
might ideally like.