Remove field that's always empty.
`jit_ir/mod.rs` and `trace_builder.rs` both have similar (but
different!) structs called `InlinedFrame`. The former is used to
transmit data between the compilation of different traces, but -- as a
comment below shows -- side-traces can never use the `args` of a frame
inlined before the start of the side trace.

This commit removes the `args` field from `jir_ir::InlinedFrame`, and
explicitly comments -- in the right place I believe! -- that when we
copy such frames over to `trace_builder` that `args` can safely be set
to the empty list.