Optimise lookup_aot_block().
perf identifies lookup_aot_block() as one of the main costs of the
tracebuilder when running the CD benchmark.

This function is very hot (called once per block in the collected
trace). In turn we call Module::funcidx() whose job it is to find a
named function's index. It does this by looping over the functions of
the module doing string comparisons.

This change adds a function index cache to avoid the looping and string
comparisons. While we are here, we can avoid some related string
conversions too.

Performance change summary (for the CD benchmark):
 - Same number of traces compiled.
 - Half the compilation time.

Typical stats change:
```
-Total Runtime: 76711711us
+Total Runtime: 76719054us
 {
-    "duration_compiling": 34.721,
-    "duration_deopting": 0.408,
-    "duration_jit_executing": 69.668,
-    "duration_outside_yk": 5.612,
-    "duration_trace_mapping": 5.427,
-    "duration_tracing": 1.014,
-    "trace_executions": 290238,
+    "duration_compiling": 15.444,
+    "duration_deopting": 0.411,
+    "duration_jit_executing": 69.906,
+    "duration_outside_yk": 5.473,
+    "duration_trace_mapping": 5.424,
+    "duration_tracing": 0.917,
+    "trace_executions": 293291,
     "traces_compiled_err": 0,
     "traces_compiled_ok": 975,
     "traces_recorded_err": 6,
     "traces_recorded_ok": 981
 }
 ```