Fix tests for non-standard cargo profiles.
I wanted to see if running the test suite with `--profile
release-with-asserts` would shake out any "easy to fix" unsoundness, but
currently running tests with a non "debug" or "release" cargo profile
will lead to linker errors like:

```
ld.lld: error: unable to find library -lykcapi
```

This is because we compute the linker path from cargo's `PROFILE`
environment, which can only contain "release" or "debug", but when you
build with a non-standard profile, the linker path is neither of those.

From the Cargo docs:

> PROFILE -- "release" for release builds, "debug" for other builds.

https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-build-scripts

(Confusingly, if you build with `--profile release-with-debug`, then
`PROFILE` will be "release")

This change fixes running the tests with non-standard cargo profiles by
computing the linker path from the target sub-directory.